Commented out what does not work and added what a unimplented Function UE does not like pure virutal functions?

+++ Source/URoboSim/Private/Controller/ControllerType/JointController/RJointTrajectoryController.cpp
@@ -95,7 +95,7 @@ bool URJointTrajectoryController::CheckTrajectoryPoint()
 
 bool URJointTrajectoryController::CheckTrajectoryGoalReached()
 {
-  if (TrajectoryPointIndex >= Trajectory.Num())
+  if ((int)TrajectoryPointIndex >= Trajectory.Num())
   {
     // State = UJointControllerState::Normal;
     SwitchToNormal();
